Add CHANGELOG generation during module build.
Update the .gemfile to source gems from https://rubygems.org to fix a
Travis/bundler warning.
Update the Modulefile summary.
Minor changes to the README.md to clarify what the module is and does
and sync with the Modulefile description.
Update the .fixtures.yml for proper YAML formatting.
Update .travis.yml to support updated Puppet versions.